+++ lib/plugins/index.js
@@ -7,7 +7,7 @@ module.exports = [
   require('./install.js'),
   require('./package/package.js'),
   require('./deploy.js'),
-  require('./invoke/invoke.js'),
+  require('./invoke.js'),
   require('./info.js'),
   require('./logs/logs.js'),
   require('./metrics/metrics.js'),
