+++ lib/plugins/invoke.js
