+++ test/unit/lib/plugins/invoke.test.js
@@ -2,8 +2,8 @@
 
 const chai = require('chai');
 const overrideEnv = require('process-utils/override-env');
-const Invoke = require('../../../../../lib/plugins/invoke/invoke');
-const Serverless = require('../../../../../lib/Serverless');
+const Invoke = require('../../../../lib/plugins/invoke');
+const Serverless = require('../../../../lib/Serverless');
 
 chai.use(require('chai-as-promised'));
